arduino_cli: Determine board details from boards.txt
To flash the arduino(s), a few details like upload speed and mcu
type need to be dermined. To ease this, we attempt to use the
boards.txt file from the arduino IDE and get the needed details.
It is reasonable to assume that an arduino user is able to determine
the model (nano, uno, etc) and the variant (atmega328, atmega168,
etc). The variant can be set via command line or in ~/.pyduin.yml
by defining a buddy. Also introduced colored output (experimental)
and reworked the part to get the dependencies into a seperate
function.