arudino_cli: Add socat proxy feature
By installing 'socat' and enabling the usage of 'socat', a proxy
for the arduino gets started on first interaction. The pins get
set up according to the pinfile on first connect. The following
connections will be made through ~/.pyduin/<ttyUSB0|ttyACM0>.tty.
To stop the proxy one has to unplug the arduino at this point in
time.