arduino: Return value for send changed
It is a bit tricky with the send function. In asynchronous scenarios
where another listener picks up the result, we would like to have a
boolean result. In CLI, we would like to wait for the result and
evaluate it. Behavior now is: No error equals True, Error equals
False, CLI mode will get str output.