Rename firmware to pyduin.cpp, cpplint firmware, bump firmware
version to 0.6.2 - I want to be able to check the quality of the
firmware, especially since some templating of the firmware seems
to be unavoidable. I do not want to maintain too much firmware
variants. Ideally, the firmware can be stitched together, so that
a greater variety of boards are able to run the firmware.