Allow firmware to be re-written to workdir before templating
Although templating runs all the time, ist has nothing to do, when
the firmware is already baked. To allow local changes to the
firmware for testing and customizing, the file is not deleted after
uploading and not overriden by default, not even in case the
available versuin is newer than the one in the buildenv.

This feature is also useful, when it comes to force the upload
of newer versions to the device.